Task/implement middleware route protection

+++ guida/src/app/api/user/route.ts
@@ -1,9 +1,8 @@
 import prisma from "@/db";
-import User from '@prisma/client'
-import { NextResponse } from 'next/server';
-import {SubmittableUser} from "@/types";
-
-
+import {NextResponse} from 'next/server';
+import {SubmittableUser, UserRole} from "@/types";
+import {auth} from "@/auth";
+import {Prisma} from "@prisma/client";
 
 /**
  * Fetch a user based on it's id, if no id is provided, all users are returned
@@ -13,22 +12,53 @@ import {SubmittableUser} from "@/types";
 export async function GET(req: Request) {
     try {
         const {searchParams} = await new URL(req.url)
-        if (searchParams.get("id") == null) {
-            const allUsers = await prisma.user.findMany({
 
-            }) 
-            return NextResponse.json(allUsers)
+        const session = await auth()
+
+        if (session) {
+            let selectStatement : Prisma.UserSelect = {};
+
+            if (session.user.roles.includes(UserRole.ADMIN) || session.user.id === searchParams.get('id')) {
+                selectStatement = {
+                    id: true,
+                    email: true,
+                    password: true,
+                    username: true,
+                    name: true,
+                    createdAt: true,
+                    updatedAt: true,
+                    roles: true,
+                }
+            } else {
+                selectStatement = {
+                    id: true,
+                    email: true,
+                    name: true,
+                    createdAt: true,
+                    roles: true,
+                }
+            }
+
+            if (searchParams.get("id") == null) {
+                const allUsers = await prisma.user.findMany({
+                    select: selectStatement
+                })
+                return NextResponse.json(allUsers)
+            }
+
+            const userid = getId(req)
+
+            if (!userid) {
+                return NextResponse.json({error: 'not a valid id'}, {status: 400})
+            }
+            const user = await prisma.user.findFirst({
+                select: selectStatement,
+                where: {id: userid}
+            })
+            return NextResponse.json(user)
+        } else {
+            return NextResponse.json({ error: 'Not allowed!' }, { status: 405 })
         }
-
-        const userid = getId(req)
-
-        if (!userid) {
-            return NextResponse.json({error: 'not a valid id'}, {status: 400})
-        }
-        const user = await prisma.user.findFirst({
-            where: {id: userid}
-        })
-        return NextResponse.json(user)
     } catch (error) {
         return NextResponse.json({ error: 'Something went wrong!' }, { status: 500 });
     }  
@@ -43,37 +73,43 @@ export async function GET(req: Request) {
 
 export async function POST(req: Request) {
     try {
-        const user : SubmittableUser = await req.json()
+        const session = await auth()
 
-        // Retrieve the Role records that match the roles provided in submittableUser
-        const roles = await prisma.role.findMany({
-            where: {
-                name: {
-                    in: user.roles,
+        if (session && session.user.roles.includes(UserRole.ADMIN)) {
+            const user : SubmittableUser = await req.json()
+
+            // Retrieve the Role records that match the roles provided in submittableUser
+            const roles = await prisma.role.findMany({
+                where: {
+                    name: {
+                        in: user.roles,
+                    },
                 },
-            },
-        });
+            });
 
-        // If roles are not found, throw an error (optional)
-        if (roles.length === 0) {
-            throw new Error("No matching roles found for user.");
-        }
-
-        if (!user.email || !user.password || !user.username) {
-            return NextResponse.json({error: "Invalid or missing attributes for user"}, {status: 400 })
-        }
-        const newUser = await prisma.user.create({
-            data: {
-                email: user.email,
-                username: user.username,
-                password: user.password,
-                name: user.name,
-                roles: {
-                    connect: roles.map((role) => ({ id: role.id })),
-                },
+            // If roles are not found, throw an error (optional)
+            if (roles.length === 0) {
+                throw new Error("No matching roles found for user.");
             }
-        })
-        return NextResponse.json(newUser, {status: 201});
+
+            if (!user.email || !user.password || !user.username) {
+                return NextResponse.json({error: "Invalid or missing attributes for user"}, {status: 400 })
+            }
+            const newUser = await prisma.user.create({
+                data: {
+                    email: user.email,
+                    username: user.username,
+                    password: user.password,
+                    name: user.name,
+                    roles: {
+                        connect: roles.map((role) => ({ id: role.id })),
+                    },
+                }
+            })
+            return NextResponse.json(newUser, {status: 201});
+        } else {
+            return NextResponse.json({ error: 'Not allowed!' }, { status: 405 })
+        }
         
     } catch (error) {
         return NextResponse.json({error: `could not create user}` + error}, {status: 500})
@@ -90,36 +126,46 @@ export async function PUT(req: Request) {
     try {
         const user : SubmittableUser = await req.json()
 
-        // Retrieve the Role records that match the roles provided in submittableUser
-        const roles = await prisma.role.findMany({
-            where: {
-                name: {
-                    in: user.roles,
-                },
-            },
-        });
+        const session = await auth()
 
-        // If roles are not found, throw an error (optional)
-        if (roles.length === 0) {
-            throw new Error("No matching roles found for user.");
-        }
+        if (session) {
+            if (session.user.roles.includes(UserRole.ADMIN) || session.user.id === user.id?.toString()) {
+                // Retrieve the Role records that match the roles provided in submittableUser
+                const roles = await prisma.role.findMany({
+                    where: {
+                        name: {
+                            in: user.roles,
+                        },
+                    },
+                });
 
-        if (Number.isNaN(Number(user.id))) {
-            return NextResponse.json({error: 'not a valid id'}, {status: 400})
-        }
-        const updatedUser = await prisma.user.update({
-            where: {id: Number(user.id)},
-            data: {
-                email: user.email,
-                username: user.username,
-                password: user.password,
-                name: user.name,
-                roles: {
-                    connect: roles.map((role) => ({ id: role.id })),
-                },
+                // If roles are not found, throw an error (optional)
+                if (roles.length === 0) {
+                    throw new Error("No matching roles found for user.");
+                }
+
+                if (Number.isNaN(Number(user.id))) {
+                    return NextResponse.json({error: 'not a valid id'}, {status: 400})
+                }
+                const updatedUser = await prisma.user.update({
+                    where: {id: Number(user.id)},
+                    data: {
+                        email: user.email,
+                        username: user.username,
+                        password: user.password,
+                        name: user.name,
+                        roles: {
+                            connect: roles.map((role) => ({ id: role.id })),
+                        },
+                    }
+                })
+                return NextResponse.json(updatedUser, {status: 201})
+            } else {
+                return NextResponse.json({ error: 'Not allowed!' }, { status: 405 })
             }
-        })
-        return NextResponse.json(updatedUser, {status: 201})
+        } else {
+            return NextResponse.json({ error: 'You must be logged in!' }, { status: 405 })
+        }
         
     } catch (error) {
         console.log(error)
@@ -139,10 +185,21 @@ export async function DELETE(req: Request) {
         if (!userid) {
             return NextResponse.json({error: 'not a valid id'}, {status: 400})
         }
-        const user = await prisma.user.delete({
-            where: {id: userid}
-        })
-        return NextResponse.json({message: "User deleted succesfully", user: user})
+
+        const session = await auth()
+
+        if (session) {
+            if (session.user.roles.includes(UserRole.ADMIN)) {
+                const user = await prisma.user.delete({
+                    where: {id: userid}
+                })
+                return NextResponse.json({message: "User deleted successfully", user: user})
+            } else {
+                return NextResponse.json({ error: 'Not allowed' }, { status: 405 })
+            }
+        } else {
+            return NextResponse.json({ error: 'You must be logged in!' }, { status: 405 })
+        }
     } catch (error) {
         return NextResponse.json({ error: 'Something went wrong! Could not delete user' }, { status: 500 });
     }
